+++ CHANGELOG.md
@@ -216,3 +216,4 @@
 - 2023-12-18: Impl bettercontact, discover, deal, main, +2 query
 - 2023-12-19: Pool leads, bettercontact, deal, conf, +3
 - 2023-12-20: Valid conf, campaign, README, discover, +2
+- 2023-12-21: Init mod, main, LICENSE, qualify, +3